requirements.txt
pandas
numpy
telegram-send
tqdm
fastparquet
plotly
openpyxl
xlrd
kaleido
pillow
statsmodels
python-dotenv
tabulate
matplotlib
seaborn
multiprocess
requests
ipykernel
localprojections
scikit-learn
pypdf
reportlab